+++ Assets/Core/_Scripts/Enemy/EnemyData.cs
@@ -7,5 +7,5 @@ public class EnemyData : ScriptableObject
     public int damage;
     public float attackCooldown;
     public float speed;
-    public int GoldOnDeath;
+    public ulong GoldOnDeath;
 }
